+++ src/main/kotlin/ca/joelathiessen/kaly2/featuredetector/SplitAndMergeFeature.kt
@@ -1,6 +1,6 @@
 package ca.joelathiessen.kaly2.featuredetector
 
-class SplitAndMergeFeature(sensorX: Double, sensorY: Double, distance: Double, angle: Double, stdDev: Double = 0.0) :
+class SplitAndMergeFeature(sensorX: Float, sensorY: Float, distance: Float, angle: Float, stdDev: Float = 0.0f) :
         Feature(sensorX, sensorY, distance, angle, stdDev) {
 
     var discardedPoints: Int = 0// the number of points along the line starting at this point that were discarded during the split and merge
